Add support to igarapes in the delivery path function. Add seasonality feature to the delivery path function.

+++ Projects/Project1/src/river.h
@@ -6,17 +6,17 @@
 class River
 {
 public:
-    River(const std::string& name = "", int maxCapacity = 0) : _name(name), _maxCapacity(maxCapacity) { }
+    River(const std::string& name = "", double maxCapacity = 0.0) : _name(name), _maxCapacity(maxCapacity) { }
 
     const std::string& GetName() const { return _name; }
-    int GetMaxCapacity()  const { return _maxCapacity; }
+    double GetMaxCapacity()  const { return _maxCapacity; }
 
     void SetName(const std::string& newName) { _name = newName; }
-	void SetMaxCapacity(int newMaxCapacity) { _maxCapacity = newMaxCapacity; }
+	void SetMaxCapacity(double newMaxCapacity) { _maxCapacity = newMaxCapacity; }
 
 private:
     std::string _name;
-    int            _maxCapacity;
+    double            _maxCapacity;
 };
 
 #endif // RIVER_H_
